classification example

+++ cs455/spark/hw4/src/main/java/Song.java
@@ -13,13 +13,13 @@ public class Song implements Serializable {
     public double duration;
     public double end_of_fade_in;
     public double energy;
-    public double key;
+    public int key;
     public double loudness;
-    public double mode;
+    public int mode;
     public double start_of_fade_out;
     public double tempo;
-    public double time_signature;
-    public double year;
+    public int time_signature;
+    public int year;
     
     public String getArtist_Terms() { return artist_terms; }
     public void setArtist_Terms(String artist_terms) { this.artist_terms = artist_terms; }
@@ -36,14 +36,14 @@ public class Song implements Serializable {
     public double getEnergy() { return energy; }
     public void setEnergy(double energy) { this.energy = energy; }
     
-    public double getKey() { return key; }
-    public void setKey(int key) { this.key = (double) key; }
+    public int getKey() { return key; }
+    public void setKey(int key) { this.key = key; }
 	
     public double getLoudness() { return loudness; }
     public void setLoudness(double loudness) { this.loudness = loudness; }
     
-    public double getMode() { return mode; }
-    public void setMode(int mode) { this.mode = (double) mode; }
+    public int getMode() { return mode; }
+    public void setMode(int mode) { this.mode = mode; }
 	
     public double getStart_Of_Fade_Out() { return this.start_of_fade_out; }
     public void setStart_Of_Fade_Out(double start_of_fade_out) { this.start_of_fade_out = start_of_fade_out; }
@@ -51,10 +51,10 @@ public class Song implements Serializable {
     public double getTempo() { return tempo; }
     public void setTempo(double tempo) { this.tempo = tempo; }
 	
-    public double getTime_Signature() { return time_signature; }
+    public int getTime_Signature() { return time_signature; }
     public void setTime_Signature(int time_signature) { this.time_signature = time_signature; }
     
-    public double getYear() { return year; }
-    public void setYear(int year) { this.year = (double) year; }    
+    public int getYear() { return year; }
+    public void setYear(int year) { this.year = year; }    
     
 }
